+++ packages/ai/core/generate-object/stream-object.ts
@@ -6,9 +6,14 @@ import {
   LanguageModelV2Usage,
   SharedV2ProviderMetadata,
 } from '@ai-sdk/provider';
-import { createIdGenerator, Schema } from '@ai-sdk/provider-utils';
+import {
+  createIdGenerator,
+  type InferSchema,
+  type Schema,
+} from '@ai-sdk/provider-utils';
 import { ServerResponse } from 'http';
-import { z } from 'zod';
+import * as z3 from 'zod/v3';
+import * as z4 from 'zod/v4/core';
 import { NoObjectGeneratedError } from '../../src/error/no-object-generated-error';
 import { createTextStreamResponse } from '../../src/text-stream/create-text-stream-response';
 import { pipeTextStreamToResponse } from '../../src/text-stream/pipe-text-stream-to-response';
@@ -25,7 +30,9 @@ import { CallSettings } from '../prompt/call-settings';
 import { convertToLanguageModelPrompt } from '../prompt/convert-to-language-model-prompt';
 import { prepareCallSettings } from '../prompt/prepare-call-settings';
 import { Prompt } from '../prompt/prompt';
+import { resolveLanguageModel } from '../prompt/resolve-language-model';
 import { standardizePrompt } from '../prompt/standardize-prompt';
+import { wrapGatewayError } from '../prompt/wrap-gateway-error';
 import { assembleOperationName } from '../telemetry/assemble-operation-name';
 import { getBaseTelemetryAttributes } from '../telemetry/get-base-telemetry-attributes';
 import { getTracer } from '../telemetry/get-tracer';
@@ -41,7 +48,6 @@ import { LanguageModelUsage } from '../types/usage';
 import { getOutputStrategy, OutputStrategy } from './output-strategy';
 import { ObjectStreamPart, StreamObjectResult } from './stream-object-result';
 import { validateObjectGenerationInput } from './validate-object-generation-input';
-import { resolveLanguageModel } from '../prompt/resolve-language-model';
 
 const originalGenerateId = createIdGenerator({ prefix: 'aiobj', size: 24 });
 
@@ -155,25 +161,19 @@ functionality that can be fully encapsulated in the provider.
 A result object for accessing the partial object stream and additional information.
  */
 export function streamObject<
-  RESULT extends SCHEMA extends z.Schema
-    ? Output extends 'array'
-      ? Array<z.infer<SCHEMA>>
-      : z.infer<SCHEMA>
-    : SCHEMA extends Schema<infer T>
-      ? Output extends 'array'
-        ? Array<T>
-        : T
-      : never,
-  SCHEMA extends z.Schema | Schema = z.Schema<JSONValue>,
-  Output extends
+  SCHEMA extends z3.Schema | z4.$ZodType | Schema = z4.$ZodType<JSONValue>,
+  OUTPUT extends
     | 'object'
     | 'array'
     | 'enum'
-    | 'no-schema' = RESULT extends string ? 'enum' : 'object',
+    | 'no-schema' = InferSchema<SCHEMA> extends string ? 'enum' : 'object',
+  RESULT = OUTPUT extends 'array'
+    ? Array<InferSchema<SCHEMA>>
+    : InferSchema<SCHEMA>,
 >(
   options: Omit<CallSettings, 'stopSequences'> &
     Prompt &
-    (Output extends 'enum'
+    (OUTPUT extends 'enum'
       ? {
           /**
 The enum values that the model should use.
@@ -182,7 +182,7 @@ The enum values that the model should use.
           mode?: 'json';
           output: 'enum';
         }
-      : Output extends 'no-schema'
+      : OUTPUT extends 'no-schema'
         ? {}
         : {
             /**
@@ -219,7 +219,7 @@ Default and recommended: 'auto' (best mode for the model).
       */
             mode?: 'auto' | 'json' | 'tool';
           }) & {
-      output?: Output;
+      output?: OUTPUT;
 
       /**
 The language model to use.
@@ -261,13 +261,13 @@ Callback that is called when the LLM response and the final object validation ar
       };
     },
 ): StreamObjectResult<
-  Output extends 'enum'
+  OUTPUT extends 'enum'
     ? string
-    : Output extends 'array'
+    : OUTPUT extends 'array'
       ? RESULT
       : DeepPartial<RESULT>,
-  Output extends 'array' ? RESULT : RESULT,
-  Output extends 'array'
+  OUTPUT extends 'array' ? RESULT : RESULT,
+  OUTPUT extends 'array'
     ? RESULT extends Array<infer U>
       ? AsyncIterableStream<U>
       : never
@@ -284,7 +284,9 @@ Callback that is called when the LLM response and the final object validation ar
     headers,
     experimental_telemetry: telemetry,
     providerOptions,
-    onError,
+    onError = ({ error }: { error: unknown }) => {
+      console.error(error);
+    },
     onFinish,
     _internal: {
       generateId = originalGenerateId,
@@ -394,7 +396,7 @@ class DefaultStreamObjectResult<PARTIAL, RESULT, ELEMENT_STREAM>
     schemaName: string | undefined;
     schemaDescription: string | undefined;
     providerOptions: ProviderOptions | undefined;
-    onError: StreamObjectOnErrorCallback | undefined;
+    onError: StreamObjectOnErrorCallback;
     onFinish: StreamObjectOnFinishCallback<RESULT> | undefined;
     generateId: () => string;
     currentDate: () => Date;
@@ -429,7 +431,7 @@ class DefaultStreamObjectResult<PARTIAL, RESULT, ELEMENT_STREAM>
         controller.enqueue(chunk);
 
         if (chunk.type === 'error') {
-          onError?.({ error: chunk.error });
+          onError({ error: wrapGatewayError(chunk.error) });
         }
       },
     });