+++ packages/ai/core/types/language-model.ts
@@ -1,4 +1,3 @@
-import { GatewayModelId } from '@ai-sdk/gateway';
 import {
   LanguageModelV2,
   LanguageModelV2CallWarning,
@@ -9,7 +8,7 @@ import {
 /**
 Language model that is used by the AI SDK Core functions.
 */
-export type LanguageModel = GatewayModelId | LanguageModelV2;
+export type LanguageModel = string | LanguageModelV2;
 
 /**
 Reason why a language model finished generating a response.